packages/cli: Require snapshot hashes in topics.
- Remove gossipTopic
- Require snapshotHash for apps with snapshots

Inside core, Canvas.initialize() now strictly requires a composite topic to be provided, in the format `${base}#${hash}` if the application is being initialized with a snapshot.

- Changes `canvas run --topic` to specify a base topic.
- The base topic will be used when saving the application to disk, and will be passed to Canvas.initialize() as `topic`.
- If an application is started with a snapshot, which can only happen if the application is started from a directory with snapshot.bin, then Canvas.initialize() will be called with `${baseTopic}#${snapshotHash}` instead.

This means messages are now signed with the `${base}#${hash}` topic and cannot be mixed between applications of different topics.

We assume that hashes are robust to collisions. A later version might hash both the contract code and topic in the same namespace scheme.